[IMP] crm: make crm.lead phone field editable
Before this commit, the user could not edit the lead phone field
when a customer was assigned. The phone field was related to the
phone of that customer.

Now, the field related to the phone of the customer is removed.
We can edit the phone field of the lead without modifying the customer
one.

task-2083470

closes odoo/odoo#38350

Signed-off-by: Thibault Delavallee (tde) <tde@openerp.com>