[FIX] website: access to themes images
Since fa9b9d699242bda5cef522e62f226215314474ee, when accessing a website
connected as portal or public user, the themes images are not showed. An
error is arose 'Sorry, you are not allowed to access documents of type
'Attachment' (ir.attachment)'.

Now, the portal and public user have access to the themes images.

opw-2082468

closes odoo/odoo#39025

X-original-commit: 9b8c390cd5526556346356f260a9cfe1b90cd510
Signed-off-by: Jorge Pinna Puissant (jpp) <jpp@odoo.com>
Co-authored-by: Lucas Perais <lpe@odoo.com>