[FIX] website_sale, website_sale_product_configurator: fix cart update
Commits c71cfb18a5f and ea9138daea5 moved things around,
"without any functional change".
They made it so that if you add twice a product to the cart,
they are added in different lines instead of merged.

There are two issues:
- the `if line_id: return lines` made sense when the post-filtering
was in the same function, since it could readily be avoided
- conversely the post-filtering was avoided in the function that was added with
the only goal to do it, except in the case where it was useless
(the case where lines is an empty recordset).

opw 2083729

closes odoo/odoo#38938

X-original-commit: c9c0a0c322f02974bc6395410ca3d7213899c7e0
Signed-off-by: Nans Lefebvre (len) <len@odoo.com>