[FIX] website: display correct website logo on SEO dialog
Currently, the website logo is not displayed on SEO dialog and instead the
company logo is being displayed.

After recent changes (see [1]), company logo and website logo are now
different and so changes were not adapted on the website.

This commit fixes the issue by adapting the changes and displaying the
correct website logo on the SEO dialog.

[1] - https://github.com/odoo/odoo/commit/ebaaeb8eeb8621b8aaa21910e80a4b51f1abc7b9

task-2083006

closes odoo/odoo#42391

X-original-commit: 9c75960e931bc77530074f60997cced4b5907d3e
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>